refactor: move endpoints to a separate file
