+++ packages/rest/src/index.ts
@@ -1,3 +1,4 @@
+export * from './endpoints.js';
 export * from './invalidBucket.js';
 export * from './manager.js';
 export * from './queue.js';
